requirements.txt
PyYAML==6.0.2
nautilus_trader==1.216.0
numpy
optuna
pandas
torch
